refresh sampling when task is succesful for all pipelines
